+++ libafl_qemu/src/arch/arm.rs
@@ -8,7 +8,7 @@ use pyo3::prelude::*;
 pub use strum_macros::EnumIter;
 pub use syscall_numbers::arm::*;
 
-use crate::{sync_exit::BackdoorArgs, CallingConvention};
+use crate::{sync_exit::ExitArgs, CallingConvention};
 
 /// Registers for the ARM instruction set.
 #[derive(IntoPrimitive, TryFromPrimitive, Debug, Clone, Copy, EnumIter)]
@@ -33,19 +33,19 @@ pub enum Regs {
     R25 = 25,
 }
 
-static BACKDOOR_ARCH_REGS: OnceLock<EnumMap<BackdoorArgs, Regs>> = OnceLock::new();
+static EXIT_ARCH_REGS: OnceLock<EnumMap<ExitArgs, Regs>> = OnceLock::new();
 
-pub fn get_backdoor_arch_regs() -> &'static EnumMap<BackdoorArgs, Regs> {
-    BACKDOOR_ARCH_REGS.get_or_init(|| {
+pub fn get_exit_arch_regs() -> &'static EnumMap<ExitArgs, Regs> {
+    EXIT_ARCH_REGS.get_or_init(|| {
         enum_map! {
-            BackdoorArgs::Ret  => Regs::R0,
-            BackdoorArgs::Cmd  => Regs::R0,
-            BackdoorArgs::Arg1 => Regs::R1,
-            BackdoorArgs::Arg2 => Regs::R2,
-            BackdoorArgs::Arg3 => Regs::R3,
-            BackdoorArgs::Arg4 => Regs::R4,
-            BackdoorArgs::Arg5 => Regs::R5,
-            BackdoorArgs::Arg6 => Regs::R6,
+            ExitArgs::Ret  => Regs::R0,
+            ExitArgs::Cmd  => Regs::R0,
+            ExitArgs::Arg1 => Regs::R1,
+            ExitArgs::Arg2 => Regs::R2,
+            ExitArgs::Arg3 => Regs::R3,
+            ExitArgs::Arg4 => Regs::R4,
+            ExitArgs::Arg5 => Regs::R5,
+            ExitArgs::Arg6 => Regs::R6,
         }
     })
 }
